refactor(api): separate prompting LLM and streaming response
