+++ api/src/llm/infrastructure/repositories/prompt-repository.js
@@ -1,5 +1,3 @@
-import { PassThrough, pipeline, Transform } from 'node:stream';
-
 import jwt from 'jsonwebtoken';
 
 import { config } from '../../../shared/config.js';
@@ -20,63 +18,9 @@ const logger = child('llm:api', { event: SCOPES.LLM });
  * @param {string} params.message
  * @param {Configuration} params.configuration
  * @param {Chat} params.chat
- * @param {Function} params.onLLMResponseReceived Callback called when LLM response has been completely retrieved. Will be called asynchronously with one parameter: the complete LLM message
- * @returns {Promise<module:stream.internal.PassThrough>}
+ * @returns {Promise<ReadableStream>}
  */
-export async function prompt({ message, configuration, chat, onLLMResponseReceived }) {
-  const readableStream = await postUserPrompt({ message, configuration, chat });
-
-  const decoder = new TextDecoder();
-  const transformFindObjects = new Transform({
-    objectMode: true,
-    transform(chunk, _encoding, callback) {
-      const objects = findObjects(decoder.decode(chunk));
-      for (const object of objects) {
-        this.push(object);
-      }
-      callback();
-    },
-  });
-
-  let completeLLMMessage = '';
-  const transformConvertObjectToEventStreamData = new Transform({
-    objectMode: true,
-    transform(chunk, _encoding, callback) {
-      const { message } = chunk;
-      if (!message) {
-        callback();
-        return;
-      }
-      completeLLMMessage += message;
-      const data = toEventStreamData(message);
-      callback(null, data);
-    },
-  });
-
-  const writableStream = new PassThrough();
-  writableStream.on('error', (err) => {
-    logger.error(`error while streaming response: ${err}`);
-  });
-  pipeline(
-    readableStream,
-    transformFindObjects,
-    transformConvertObjectToEventStreamData,
-    writableStream,
-    async (err) => {
-      if (err) {
-        logger.error(`error in pipeline: ${err}`);
-        if (!writableStream.closed && !writableStream.errored) {
-          writableStream.end('Error while streaming response from LLM');
-        }
-      } else {
-        await onLLMResponseReceived(completeLLMMessage);
-      }
-    },
-  );
-  return writableStream;
-}
-
-async function postUserPrompt({ message, configuration, chat }) {
+export async function prompt({ message, configuration, chat }) {
   const messagesToForward = chat.messages.slice(-configuration.historySize).map(toHistoryMessage);
   const payload = JSON.stringify({
     prompt: message,
@@ -128,22 +72,4 @@ function toHistoryMessage(message) {
     content: message.content,
     role: message.isFromUser ? 'user' : 'assistant',
   };
-}
-
-export function toEventStreamData(message) {
-  const formattedMessage = message.replaceAll('\n', '\ndata: ');
-  return `data: ${formattedMessage}\n\n`;
-}
-
-export function findObjects(str) {
-  const objects = [];
-  while (str.length > 0) {
-    const [numberAsStr, ...otherParts] = str.split(':');
-    const objectLength = parseInt(numberAsStr);
-    const strLeft = otherParts.join(':');
-    objects.push(strLeft.slice(0, objectLength));
-    str = strLeft.slice(objectLength);
-  }
-
-  return objects.map((obj) => JSON.parse(obj));
 }
